titanobova-django/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.21.0
django-filter==23.4
django-guardian==2.4.0
djangorestframework-simplejwt==5.3.2
cryptography==41.0.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
requests==2.31.0
stripe==7.2.0
python-decouple==3.8
email-validator==2.1.0
django-ratelimit==4.1.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
sendgrid==6.10.0
django-storages==1.14.2
boto3==1.33.2
whitenoise==6.6.0
django-cacheops==6.2
sentry-sdk==1.38.0
django-extensions==3.2.3
black==23.12.0
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
